+++ amiga-mui/audioselectgroupclass.c
@@ -98,7 +98,7 @@ STATIC ULONG AudioSelectGroup_New(struct IClass *cl,Object *obj,struct opSet *ms
 	data->file_string = file_string;
 	data->play_button = play_button;
 
-	DoMethod(play_button, MUIM_Notify, MUIA_Pressed, FALSE, App, 4, MUIM_CallHook, &hook_standard, AudioSelectGroup_Play, data);
+	DoMethod(play_button, MUIM_Notify, MUIA_Pressed, FALSE, (ULONG)App, 4, MUIM_CallHook, (ULONG)&hook_standard, (ULONG)AudioSelectGroup_Play, (ULONG)data);
 
 	return (ULONG)obj;
 }
